requirements.txt
requests
sanic[ext]
tortoise-orm[asyncpg]
aerich